[modify] rename struct.
