+++ PyConJP2016/Model/Place.swift
@@ -1,5 +1,5 @@
 //
-//  LevelModel.swift
+//  PlaceModel.swift
 //  PyConJP2016
 //
 //  Created by Yutaro Muta on 2016/02/22.
@@ -8,22 +8,14 @@
 
 import UIKit
 
-//class LevelModel: NSObject {
-//    
-//    var id: Int = 0
-//    var name: String = ""
-//    
-//    init(dictionary: [String : AnyObject]) {
-//        id = dictionary["id"] as! Int
-//        name = dictionary["name"] as! String
-//    }
-//    
-//}
-
-struct Level {
+struct Place {
     
     let id: Int
     let name: String
+
+}
+
+extension Place {
     
     init(dictionary: [String : AnyObject]) {
         id = dictionary["id"] as? Int ?? 0
